Use open_port's option eof instead
This makes the code more sane since we don't have to handle the
EXIT-message which might or might not come (depending on if it's and
API-call or AMQP connection).
Also, on Tim Watsons (indirect) suggestion, make the code more
erlang-like (and probably more/better tail recursion?) by using an
accumulator.